final preparation and aligment to extract web into separate project
