+++ toloka2MediaServerWeb/app.py
@@ -1,5 +1,5 @@
 import logging
-from flask import Flask, json, jsonify, request, render_template, redirect, url_for, session, Response
+from flask import Flask, jsonify, request, render_template, session, Response
 import requests
 import sys
